test: split rate-limit resolver test to satisfy phpmd TooManyMethods
ApiRateLimitRequestResolverLimitersTest had 28 methods, exceeding the
25-method cap in phpmd.tests.xml. Move the 10 auth/recovery/GraphQL-flow
tests into a new ApiRateLimitRequestResolverAuthLimitersTest, leaving 18
in the original. No assertions changed.
